# Speed up dictionary iteration by reducing branchmisses