@JeffBezanson 2019-03-04
Clever!

I don't see how this removes the need for `idxfloor` --- aren't there still cases where repeatedly iterating could take asymptotically longer?