@chethega 2019-03-04
Re skipping idxfloor:

The only case when idxfloor is helpful is if it is very large, i.e. if we could skip ahead by a large amount. In the old code, long stretches of empty or full slots are very cheap: The branch predictor catches on, and skips all these guys.

But more importantly: Filled slots are randomly distributed (unless our hash function failed). Suppose k out of N slots are full; we have ~k gaps of average size ~N/k. The idxfloor permits us to skip a single gap. If k>10, then this is a rather meaningless optimization. The only really good case is k=1 and N large, i.e. a single entry in a giant dictionary, and idxfloor points at that entry. This can only happen if the dictionary is the remnant of a much larger dict, where all except for a single entry were removed, and the dict got rehashed afterwards (otherwise idxfloor does not point to our lonely entry).

On the other hand, idxfloor complicates the management of the dictionary state: Every operation needs to make sure that it adjusts idxfloor as appropriate. 

But if we do get to change the layout, then it would be quite nice to compress `slots` further: Two bit per slot, similar to bitvector. The advantage is again to reduce branch-misses: Currently we pay 1 miss / 8 slots. On my machine (predictable 2ghz, 20 cycle branch miss penalty) this makes for a total of about `2.5 * ρ` cycles per item, where `ρ` is the inverse density ( `2^14/5000 = 3.3` for `d2`). That could go down to `0.625 * ρ` cycles per item. 

That is not insignificant: The `collect(d2)` ran in 10 cycles/item at an inverse density of `3.3`, i.e. 8/10 cycles per item go into the branch miss. If we compressed further, we would pay 2 cycles/item in branch misses; it seems realistic that another factor 2 is available. Downside is that this would slightly complicate indexing code; and it is only possible in a major release because too many people rely on dictionary internals.