@chethega 2019-03-07
Ok, `idxfloor` is essential: I forgot about the `pop(d::Dict)` and `pop!(d::Dict)` operations.