@vtjnash 2023-10-31
This seems a step towards SwissDict, which possibly we should just import and make the default (from https://github.com/JuliaCollections/DataStructures.jl/blob/master/src/swiss_dict.jl)